feat: Update compliance workflow and data models with new output classes and agent name handling
