+++ src/app/chainlit.md
@@ -1,73 +1,65 @@
-
-# 🤖 STU Copilot
+## 🤖 STU Copilot
 
 Welcome to **STU Copilot**! 🚀
 
 > STU Copilot is your intelligent assistant for searching, retrieving, and interacting with a wide range of technical resources. It leverages advanced AI agents to help you find documentation, code, presentations, blog posts, and more, all from a single interface. Whether you're a developer, engineer, or technical writer, STU Copilot streamlines your workflow and boosts productivity.
 
-## ⚙️ Available Tools
+### ⚙️ Available Tools
 
 STU Copilot provides the following specialized tools, each powered by a dedicated agent:
 
+---
 
-### 📄 Microsoft Docs
-- **Description:** Search Microsoft documentation for technical resources, guides, and API references.
-- **How to use:** Use the "Microsoft Docs" command to query official Microsoft docs.
-- **Example queries:**
-	- "Find the documentation for Azure Functions triggers."
-	- "How do I use Microsoft Graph API?"
+#### 📄 Microsoft Docs
 
+-   **Description:** Search Microsoft documentation for technical resources, guides, and API references.
+-   **How to use:** Use the "Microsoft Docs" command to query official Microsoft docs.
+-   **Example queries:**
+    -   "Find the documentation for Azure Functions triggers."
+    -   "How do I use Microsoft Graph API?"
+---
+#### 📄 GitHub Docs
 
-### 📄 GitHub Docs
-- **Description:** Search GitHub documentation for platform guides, API docs, and best practices.
-- **How to use:** Use the "GitHub Docs" command to find relevant GitHub documentation.
-- **Example queries:**
-	- "Show me how to create a GitHub Actions workflow."
-	- "What is GitHub Codespaces?"
+-   **Description:** Search GitHub documentation for platform guides, API docs, and best practices.
+-   **How to use:** Use the "GitHub Docs" command to find relevant GitHub documentation.
+-   **Example queries:**
+    -   "Show me how to create a GitHub Actions workflow."
+    -   "What is GitHub Codespaces?"
+---
+#### 🐙 GitHub
 
+-   **Description:** Search for GitHub repositories, code, and project information.
+-   **How to use:** Use the "GitHub" command to look up repositories or code snippets.
+-   **Example queries:**
+    -   "Find Python projects for data science."
+    -   "Show me the README for microsoft/TypeScript."
+---
+#### 📊 Seismic Presentations
 
-### 🐙 GitHub
-- **Description:** Search for GitHub repositories, code, and project information.
-- **How to use:** Use the "GitHub" command to look up repositories or code snippets.
-- **Example queries:**
-	- "Find Python projects for data science."
-	- "Show me the README for microsoft/TypeScript."
+-   **Description:** Search for Seismic content, including presentations and sales materials.
+-   **How to use:** Use the "Seismic Presentations" command to retrieve Seismic assets.
+-   **Example queries:**
+    -   "Find the latest sales enablement presentation."
+    -   "Show Seismic content about product launches."
+---
+#### 📰 Blog Posts
 
+-   **Description:** Search for technical blog posts and articles across curated sources.
+-   **How to use:** Use the "Blog Posts" command to discover relevant blog content.
+-   **Example queries:**
+    -   "Recent blog posts about Azure AI."
+    -   "Best practices for Python packaging."
+---
+#### 📄 AWS Documentation
 
-### 📊 Seismic Presentations
-- **Description:** Search for Seismic content, including presentations and sales materials.
-- **How to use:** Use the "Seismic Presentations" command to retrieve Seismic assets.
-- **Example queries:**
-	- "Find the latest sales enablement presentation."
-	- "Show Seismic content about product launches."
-
-
-### 📰 Blog Posts
-- **Description:** Search for technical blog posts and articles across curated sources.
-- **How to use:** Use the "Blog Posts" command to discover relevant blog content.
-- **Example queries:**
-	- "Recent blog posts about Azure AI."
-	- "Best practices for Python packaging."
-
-
-### 🔎 Bing Search
-- **Description:** Search the web using Bing for up-to-date information and resources.
-- **How to use:** Use the "Bing Search" command for general web queries.
-- **Example queries:**
-	- "Latest news on cloud computing."
-	- "What is Retrieval-Augmented Generation?"
-
-
-### 📄 AWS Documentation
-- **Description:** Search AWS documentation for cloud service guides, API references, and tutorials.
-- **How to use:** Use the "AWS Documentation" command to access AWS docs.
-- **Example queries:**
-	- "How to configure AWS Lambda triggers."
-	- "Find the documentation for Amazon S3 lifecycle policies."
+-   **Description:** Search AWS documentation for cloud service guides, API references, and tutorials.
+-   **How to use:** Use the "AWS Documentation" command to access AWS docs.
+-   **Example queries:**
+    -   "How to configure AWS Lambda triggers."
+    -   "Find the documentation for Amazon S3 lifecycle policies."
 
 ---
 
-
 Each tool can be accessed via its command in the chat interface. Persistent tools remain available throughout your session, while others are context-specific. Use the appropriate command or example query to activate the agent best suited to your needs.
 
-For further assistance or to extend STU Copilot, refer to the project documentation or contact the development team.
+🤝 For further assistance or to extend STU Copilot, please contact [Hossein Zahed](https://github.com/HosseinZahed/)
